Исправлены замечания, добавлен лог выполнения.
