+++ RequestAndResponse.log
@@ -1,0 +1,102 @@
+# Проверка лимита:
+- http://localhost:8991/v1/api/payments/check-limit
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 10
+}
+- тело ответа:
+{
+    "status": true
+}
+
+# Блокировка суммы (сначала на 1 руб, затем на 5000 руб)
+- http://localhost:8991/v1/api/payments/block-amount
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 5000
+}
+- тело ответа:
+{
+    "status": true
+}
+# Подтверждение платежа (должна быть ошибка, сумма больше чем было заблокировано)
+- http://localhost:8991/v1/api/payments/confirm-amount
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 9002
+}
+- тело ответа:
+{
+    "status": "BAD_REQUEST",
+    "message": "Сумма для подтверждения превышает сумму заблокированных средств"
+}
+# Подтверждение платежа (сумма меньше, чем заблокировано)
+- http://localhost:8991/v1/api/payments/confirm-amount
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 4000
+}
+- тело ответа:
+{
+    "status": true
+}
+- строка в таблице postgres:
+id	userid	limit_summa	block_amount
+1	1	6000.0	1001.0
+# Возврат средств (сначала ошибка, возвращаем больше, чем заблокировано)
+- http://localhost:8991/v1/api/payments/rollback
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 2000
+}
+- тело ответа:
+{
+    "status": "BAD_REQUEST",
+    "message": "Восстановить лимит невозможно. Сумма заблокированных средств меньше суммы платежа"
+}
+# Возврат средств (ошибки нет, сумма возврата меньше, чем залокировано в сервисе)
+- http://localhost:8991/v1/api/payments/rollback
+- тело запроса:
+{
+    "userId": 1,
+    "amount": 500
+}
+- тело ответа:
+{
+    "status": true
+}
+- строка в таблице postgres:
+id	userid	limit_summa	block_amount
+1	1	6500.0	501.0
+
+# Проверка работы планировщика (запуск передвинул на 18:09)
+- содержимое таблицы до запуска задания:
+id	userid	limit_summa	block_amount
+2	2	10000.0	0.0
+3	3	10000.0	0.0
+4	4	10000.0	0.0
+5	5	10000.0	0.0
+6	6	10000.0	0.0
+7	7	10000.0	0.0
+8	8	10000.0	0.0
+9	9	10000.0	0.0
+10	10	10000.0	0.0
+1	1	6500.0	501.0
+
+- содержимое таблицы после запуска задания:
+id	userid	limit_summa	block_amount
+2	2	10000.0	0.0
+3	3	10000.0	0.0
+4	4	10000.0	0.0
+5	5	10000.0	0.0
+6	6	10000.0	0.0
+7	7	10000.0	0.0
+8	8	10000.0	0.0
+9	9	10000.0	0.0
+10	10	10000.0	0.0
+1	1	10000.0	501.0
