+++ src/main/java/ru/vtb/javaPro/config/LimitConfig.java
@@ -1,10 +1,11 @@
 package ru.vtb.javaPro.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import ru.vtb.javaPro.config.properties.LimitProperties;
 
-@Component
+@Configuration
 @EnableConfigurationProperties(LimitProperties.class)
 public class LimitConfig {
     private final LimitProperties limitProperties;
